requirements.txt
apache-airflow==2.3.0
pendulum==2.1.2
psycopg2-binary==2.9.3
pandas==1.3.5
SQLAlchemy==1.4.9